Whee: pointer enter and leave events synthesized from move events.
Not from press/release yet, those would likely be done only for finger
events with the assumption that pointer events with mouse as a source
generate enough move events that it's very unlikely that a press happens
before a move on given target.

And I need to find & charge my pen to check how it works for that one.