Whee: add a stress test app.
Creates a million of nodes and data attachments and churns through them
every frame. At this point the whole update including GPU data setup and
upload takes about 46 ms for a million of nodes in RelWithDebInfo, or
39 with -O3 and -march=native on an AVX512 system. Which is a bit more
than I'd like, but a million of *live* nodes is a very rare scenario.

The stress test app will be gradually extended for more real-life
scenarios once culling and other features get added.