Whee: there's no use for a layer to request a clean() on itself.
The LayerState::NeedsClean flag was useful in the initial implementation
where attachment of data to nodes was managed by the UserInterface. But
when the node attachments were moved to the layer, the clean() changed
to take node handle generations instead of a BitArrayView with concrete
IDs, and it now constructs the BitArrayView internally by comparing the
generations with generations of node attachments.

Which means there's no longer any way to call clean() to remove an
arbitrary data ID, only by removing a whole node and all data attached
to it, so there's no way the NeedsClean state flag can be accounted for.
Frankly, it didn't really do anything before either -- there was nothing
tracking what IDs remove() was called on, so the BitArrayView passed to
clean() never contained those, so even if the doClean() implementation
might have wanted to do something extra for the now-removed IDs, it had
no way to know which those were.

The only way to know and perform custom cleanup was to override remove()
on the concrete layer subclass, and just ignoring the NeedsClean flag,
such as what EventLayer did. But that brought another problem where
application code could inadvertently call the base remove() API, leading
to the subclass not doing the needed cleanup. One solution would be to
make it virtual and have it call some doRemove() internally, but that'd
go against the policy of having a ~constant amount of virtual calls per
frame. Now, with a single top-level node hierarchy, it's one doDraw()
and possibly doClean() / doUpdate() per layer per frame, doRemove()
would add a virtual call for every data that gets removed, so possibly
*thousands* of additional virtual calls that are impossible to optimize
away.

So instead remove() is now made protected, and a particular layer
implementation then needs to make it public and optionally add its own
handling to it. That makes it impossible to call only the base
implementation by accident. The NeedsClean flag and the corresponding
UserInterfaceState::NeedsDataClean flag is now removed as it has no
practical use.

And given that remove() is protected, create() should be as well for
balance -- to avoid the base implementation being accidentally called
leading to derived state not created correctly, such as very easily by
using `ui.layer(handle).create()`.