Whee: decouple uniforms from styles in BaseLayer.
With the recent introduction of padding it became clear that there will
definitely be more styles than actually unique uniform buffer entries
(for example, a button with an icon + text would have different
text padding from a button with just a text, which would be also
different from a button with just an icon), and forcing the style to
duplicate uniform values this way isn't nice at all given the UBO
size limits are just 64k.

Thus now there's a style index, used the same way as before for data
creation, style transition etc., but there's additionally a possibility
to set up the BaseLayer::Shared with a style uniform count differing
from the actual style count, in which case an additional mapping
from style index to an uniform index needs to be provided.

In order to make it clear what goes into an uniform and what is just a
CPU-side value that doesn't go to the shader, the BaseLayerStyleCommon
and BaseLayerStyleItem struct were renamed to
BaseLayerCommonStyleUniform and BaseLayerStyleUniform. They'll continue
to be packed in the same special way as before, and all CPU-side values
will be supplied to setStyle() via separate loose strided array views to
allow them to be in any form in the style definition, or even be omitted
completely, such as in case of the padding values.

For simple / "lazy" use cases it's still possible to have style and
uniform count the same. Then the mapping between the two is implicit.