Whee: don't multiply BaseLayer outline color with background.
Seemed like a good idea at first, to be able to have a gradient on the
outline as well without having to specify a fourth color. Unfortunately
this bright idea fails if the background is very dark (then one has to
make the outline color channels larger than 1 to make it actually show
at all, very intuitive indeed). And the worst is when the background is
something like #00000033, where the outline just cannot mathematically
be anything else than #000000 as well.

I might reintroduce this later as a shader compile flag or something
with the default behavior being no multiplication. The gradient
application could then be done with just applying a difference (or
luminance difference) between the top and bottom background color, that
would work even if the background channels are zero.