Whee: split draw and event data ordering.
Right now in isolation the change doesn't make sense as the operation of
counting data attached to visible nodes is now performed twice for
seemingly no reason. But that will change with the introduction of nodes
with disabled events, as in that case the masks will differ between the
two.