Whee: base interface for renderers and a RendererGL implementation.
At the moment the renderer only deals with setting up appropriate
GPU state such as blending or scissor (and even then it's a ton of
code). The base and text layers aren't updated to rely on it yet either,
that's done in a separate commit to not make those changes drowned in
all the other scaffolding code.

Subsequently the renderers will also maintain custom framebuffers for
various compositing operations implemented by layers themselves.