Whee: scaffolding for compositing layers.
Just the base interface, not implemented in RendererGL and neither in
any compositing feature in any layer yet.

Also, the compositing API is made in a way that allows the
implementation to optimize and perform the operation only on the
rectangle that's actually covered with any nodes, but so far that's not
used, the UI always passes the whole area there for now. Will be done in
subsequent optimization passes.