Whee: include the actual window area in the culling process.
Without this, it'd happily draw content that's completely outside the
window area, including things like

    glScissor(x = -284, y = 60, width = 284, height = 680)

When I was doing the original implementation, it just didn't occur to me
that this would be a problem (like, why would anyone have nodes moved
out of the window area). Turns out it's rather common, for various
things like pop-out menus on touch displays, and the library should do
the right thing by default and cull them if not visible, instead of
either causing the application to draw them always for no reason or
requiring it to explicitly handle that case by flipping a Hidden flag or
removing given node tree from the top-level order.

This now also means that calls to setSize() have to update the cull
state, in addition to updating renderer framebuffer size. Otherwise
making the window bigger wouldn't show the nodes that were hidden
before, and conversely making it smaller would cause them to still be
drawn even if not needed.