Whee: use the actual clipped node rects for compositing.
Right now it's exactly matching the set of data that's being drawn, so
if two data get assigned to the same node, it'll be present twice.
That's something it should eventually filter away, however with the blur
shader there's also expansion for the radius going on, which may cause
the rects to overlap as well, so another solution for that needs to be
found. Stencil buffer?