Whee: more thoroughly test partial update() calls with event handling.
So far the explicit ui.update() calls were ensuring that no event
handlers forget to update the internal state before doing anything with
it, but the opposite needs to be tested as well -- that the internals
get updated to a consistent state no matter which UserInterfaceState
flags are set.

For the AbstractLayer, Layouter and Animator update(), advance(), draw()
etc. this was verified thoroughly enough in
AbstractUserInterfaceTest::state(), but event-related state which isn't
really passed anywhere and just traversed internally so far lacked this.
And I made a mistake when updating it and only one of the event tests as
a result, and only by accident. So better have that pinned down better.