Whee: put reusable stuff from BaseLayer into AbstractVisualLayer.
Realized that a lot of this would otherwise need to get duplicated (and
repeatedly covered with tests and docs) for the TextLayer and others.
Not nice.

This doesn't really make anything worse or more complicated as the
internal state split was already done between BaseLayer and BaseLayerGL
before. The only difference is that the AbstractVisualLayer internal
state struct now contains a StridedArrayView for the style indices that
has to be kept up-to-date from the derived implementation. It could also
have its own allocation that it would independently grow but this is
easier.