Whee: make alignment a part of TextLayer style.
And TextProperties only *optionally* override it. Otherwise it's
impossible to have the style define different alignment for certain
elements (such as dialogs having the title either centered or on the
left), and even without that an actual widget implementation would have
to copy and modify the TextProperties in order to supply a different
alignment, together with no way of knowing whether the passed alignment
was the default or was what the application code wanted to override.