Whee: allow setting a text layer style with null font handles.
Given it's possible to call create(), setText() etc. with no style set
at all if a font handle is passed in the TextProperties instead, I don't
see why setStyle() should restrict the font handles to be always
non-null.

This removes the restriction, making it then behave basically the same
as if a style wasn't set yet. Which also allows the styles to be defined
with more flexibility, for example a style for icons that only defines
how they should look, but expects the application to supply the actual
icon font data itself.

A similar logic will apply for the upcoming dynamic styles as well --
those will have all fonts set to FontHandle::Null by default, as there's
no other way, and if the application code doesn't set the dynamic style
data, create() will have to supply a font directly.