Whee: introduce TextLayer editing styles for cursor and selection.
I briefly considered trying to render them together with the text with
some sort of a specially crafted texture piece in the glyph cache, but
ultimately decided against that as it'd significantly limit the
styleability, which is one of the goals of this library.

Those are thus rendered in a separate draw call. Currently, the
assumption is that there won't be too many visible selections and thus
the draw calls are interleaved for each clip rect. If that eventually
would stop being the case, a faster way to draw would be all editing
quads first, going through all clip rects, and then all text, going
through all clip rects again. Or maybe I just perform the clipping in
the data directly instead of fiddling with scissor state.

The editing styles have a configurable padding on each side, allowing
the style to expand / shrink both the selection and the cursor
rectangles for fine-tuned appearance, and there's also an ability to
round the corners. Right now, the selection and cursor is always drawn
*below* the actual text. That seems sufficient, but drawing the cursor
on top of the text (and possibly making it invert colors) isn't out of
question, although that'd mean splitting the draw calls even further.

For cursor placement it uses the byte-to-glyph mapping APIS from the
Text library. Right now this mapping is O(n) and done at every update,
if it becomes a perf bottleneck, I'll do it on every cursor adjustment
instead, and cache the position. It also doesn't handle ligatures and
merged / reordered glyphs yet in any way, so e.g. moving a cursor over
the ff ligature will get stuck on one side and then jumps on the other
side, never being placed in the middle. Not sure what to do there
besides telling users to disable kerning for editable text.