Whee: skip uniform blending in {Base,Text}LayerAnimator if the IDs match.
Originally I assumed the uniform would change always and thus performed
uniform buffer upload always. But that's not always the case, it could
be for example just the padding changing, or, more importantly, in case
of the text editing styles, it could be just the editing uniforms, or
the base uniforms, or just editing uniform paddings, ... and uploading
all of them all the time doesn't seem wise if it's really just for a
blinking cursor.

Like before however, comparing all uniform values would get pretty nasty
pretty soon, so what this is doing is just comparing the ID, with an
assumption that style authoring was done efficiently, with uniforms
deduplicated. Uniforms with the same data but different IDs will still
trigger a reupload every time tho.

(Yes, the two animator implementations feel like a silly copypaste of
each other, but that's going to change now, with the introduction of
text editing styles.)