Whee: expect style uniform indices to be in bounds.
Originally I thought "meh, the GPU takes care of that" and didn't add an
assert, but it's damn annoying to debug. especially if an out-of-bounds
index gives the same black screen of death as about any other rendering
issue, and for style animators out-of-bounds indices cause strange
assertions deep inside the code, making it look like it's a library bug
and not user error.

So there you go, assertions for all those. I might eventually downgrade
them to debug-only, but so far they're present always.