Whee: introduce a concept of a focused node.
Together with NodeFlag::Focusable to mark nodes as being able to receive
focus (because it wouldn't make sense on just any), implicit handling of
focus in ui.pointerPressEvent(), explicit focus via ui.focusEvent() and
focusEvent() and blurEvent() interfaces on the layer.

So far for focus via a pointer press there's no fallback done, so
clicking on a non-focusable child of a focusable node will do nothing.
That may get implemented at some point later, for now I want the basics
done at least. Similarly, this is not yet exposed in AbstractVisualLayer
either.

This also expands on the VisibilityLostEvent, causing it being called
from ui.update() also if a focused node is no longer focusable. Instead
of a blur event, because for that one it's important to guarantee that
it's always called with doUpdate() before, not with sometimes before and
sometimes after. For proper style transitioning it thus needs to know
whether the no-longer-focusable node is still hovered or pressed.