Whee: allow putting non-root nodes into top-level order.
This is to enable use cases like showing tooltips positioned relative to
their parent nodes, dropdowns, menus or even the rare cases where nodes
*need* to mutually overlap and be drawn in correct order.

The support is recursive, meaning that children of a nested top-level
node can be themselves made top-level again (think a submenu opening).
It balooned up a bit because ultimately it made sense to enforce the
nested top-level nodes to be always on top of their parents, and also if
a top-level node changes its order all its child nodes go alongside it.
Which is both a convenient feature from the application perspective and
avoids a dedicated step for calculating absolute offsets (and
eventually also clip rects, NoEvent/Disabled flag propagation) because
this way the parent value is always known when the algorithm gets to the
child.