Whee: make layouters aware of UI size and its change as well.
For some reason this didn't occur to me, but is clearly essential to be
able to layout things relative to the UI itself without having to
manually propagate the size changes to each layout.

Originally I thought I'd upgrade ui.setSize() to trigger also
NeedsLayoutUpdate, but when updating the test cases I realized it may be
excessive and not always needed. For example if a particular layouter is
used only inside some (infinite) scroll area, triggering its update
would be needless. So instead layouter's doSetSize() implementation is
expected to call setNeedsUpdate() if it actually contains anything that
is UI-size-dependent. If not, then only the clip state gets updated.