Whee: require layer.setSize() to be called before layer.update().
This was already required by the AbstractUserInterface::update() itself,
but as many tests call update() directly on layers, this check was
sidestepped. Then BaseLayer added its own assert (which was impossible
to get triggered by regular user code because the AbstractUserInterface
one would fire first), the assert didn't really work if the blur scale
vector was really tiny which was an annoying bug, and doing it this way
would mean each layer implementation would eventually need to add the
same check. Which doesn't scale.

So it's now done by the AbstractLayer base directly, which *may* be
annoying for some tests, but definitely is cleaner and less
accident-prone.